fix race condition
A boolean variable in `Raytracer` was used to notify the tracer
to stop, but it was not atomic, which is a race, as there was no
locking.

The variable has been changed from `bool` to `boost::atomic<bool>`,
not `std::atomic<bool>`, because msvc10 does not have `std::atomic`
support. This requires Boost 1.53.