remove xor with random seed
Every new Monte Carlo unit was initialised with a seed xor a random
number computed by the previous unit. This is silly, just using a
random number from the previous unit suffices. Note that the number
produced is not the initial state of the random generator, so by
seeding it into the next generator, the next generator will not
produce the same values, because it will start with a different
internal state.